+++ SourceryTests/Generating/SwiftTemplateSpecs.swift
@@ -24,7 +24,7 @@ class SwiftTemplateTests: QuickSpec {
             let expectedResult = try? (Stubs.resultDirectory + Path("Basic.swift")).read(.utf8)
 
             it("generates correct output") {
-                expect { try Sourcery(cacheDisabled: true).processFiles(.sources([Stubs.sourceDirectory]), usingTemplates: [templatePath], output: outputDir) }.toNot(throwError())
+                expect { try Sourcery(cacheDisabled: true).processFiles(.sources(Paths(include: [Stubs.sourceDirectory])), usingTemplates: Paths(include: [templatePath]), output: outputDir) }.toNot(throwError())
 
                 let result = (try? (outputDir + Sourcery().generatedPath(for: templatePath)).read(.utf8))
                 expect(result).to(equal(expectedResult))